Reorganized code dealing with the command class
